Implemented the skeleton for interrupt servicing
Right now, support interrupts in the simulator is... strange.

We have to implement the functionality of triggerring the interrupt but also make sure we don't corrupt the current execution state that the simulator is running. On an actual processor, the interrupt mechanism is transparent and does a bunch of extra work for you. But in the case of the simulator, we need to implement that underlying transparent mechanism. It is not as simple as performing a call because we are still emulating cycles so we would need to preserve the current state of the emulator through a resumption record. What I don't know is how that would look to be honest.

Even if we were able to exploit the call stack, it would also be a problem if we smashed the stack. It requires extra logic to handle at this point.